gdallocationinfo: add target for windows; fix windows test
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@21187 f0d54148-0727-0410-94bb-9a71ac55c965

+++ apps/gdallocationinfo.cpp
@@ -364,8 +364,8 @@ int main( int argc, char ** argv )
         }
         else
         {
-            iPixel = floor(dfGeoX);
-            iLine  = floor(dfGeoY);
+            iPixel = (int) floor(dfGeoX);
+            iLine  = (int) floor(dfGeoY);
         }
 
     /* -------------------------------------------------------------------- */
